Properly produce JDBC non-shaded jar
This introduces a new module that publishes unshaded JDBC JAR along with a POM
containing transitive dependencies and the fat JAR along with a POM
that doesn't contain transitive dependencies. For backward compatibility,
the fat JAR is named `trino-jdbc` while unshaded is `trino-jdbc-core`.
