+++ client/trino-jdbc-core/src/main/java/io/trino/jdbc/InMemoryTrinoResultSet.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.jdbc;
+
+import io.trino.client.Column;
+
+import java.sql.SQLException;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+class InMemoryTrinoResultSet
+        extends AbstractTrinoResultSet
+{
+    private final AtomicBoolean closed = new AtomicBoolean();
+
+    public InMemoryTrinoResultSet(List<Column> columns, List<List<Object>> results)
+    {
+        super(Optional.empty(), columns, results.iterator());
+    }
+
+    @Override
+    public void close()
+            throws SQLException
+    {
+        closed.set(true);
+    }
+
+    @Override
+    public boolean isClosed()
+            throws SQLException
+    {
+        return closed.get();
+    }
+}
