+++ client/trino-jdbc-core/src/main/java/io/trino/jdbc/ObjectCasts.java
