+++ client/trino-jdbc-core/src/main/java/io/trino/jdbc/TrinoIntervalDayTime.java
