+++ client/trino-jdbc-core/src/main/resources/io/trino/jdbc/driver.properties
